Snowman Render
Render snowman parts

+++ src/main/java/com/zapcloudstudios/utils/draw/BoxDraw.java
@@ -7,37 +7,37 @@ public abstract class BoxDraw extends SimpleDraw
 	protected int width;
 	protected int height;
 	protected int length;
-	
+
 	protected boolean inside = false;
-	
+
 	public BoxDraw()
 	{
 		super();
 	}
-	
+
 	public void faceOut()
 	{
 		this.inside = false;
 	}
-	
+
 	public void faceIn()
 	{
 		this.inside = true;
 	}
-	
-	public void cube(int x, int y, int z, int width, int height, int length)
+
+	public void cube(float x, float y, float z, int width, int height, int length)
 	{
 		this.setPos(x, y, z);
 		this.cube(width, height, length);
 	}
-	
+
 	public void cube(int width, int height, int length)
 	{
 		this.width = width;
 		this.height = height;
 		this.length = length;
 	}
-	
+
 	@Override
 	public void forwardTo(DrawBase to)
 	{
@@ -51,7 +51,7 @@ public abstract class BoxDraw extends SimpleDraw
 			boxto.inside = this.inside;
 		}
 	}
-	
+
 	public void drawSide(EnumDirection dir)
 	{
 		switch (dir)
@@ -76,19 +76,19 @@ public abstract class BoxDraw extends SimpleDraw
 				break;
 		}
 	}
-	
+
 	public abstract void XUp();
-	
+
 	public abstract void XDown();
-	
+
 	public abstract void YUp();
-	
+
 	public abstract void YDown();
-	
+
 	public abstract void ZUp();
-	
+
 	public abstract void ZDown();
-	
+
 	public void drawAll()
 	{
 		this.XUp();
@@ -98,7 +98,7 @@ public abstract class BoxDraw extends SimpleDraw
 		this.ZUp();
 		this.ZDown();
 	}
-	
+
 	public void drawSidesSameTexture()
 	{
 		this.XUp();
@@ -106,7 +106,7 @@ public abstract class BoxDraw extends SimpleDraw
 		this.ZUp();
 		this.ZDown();
 	}
-	
+
 	public void drawSidesGroupedTexture()
 	{
 		int u = this.textureU;
@@ -121,7 +121,7 @@ public abstract class BoxDraw extends SimpleDraw
 		this.textureU = u;
 		this.textureV = v;
 	}
-	
+
 	public void drawAllNormalTextureShape()
 	{
 		int u = this.textureU;
@@ -137,7 +137,7 @@ public abstract class BoxDraw extends SimpleDraw
 		this.textureU = u;
 		this.textureV = v;
 	}
-	
+
 	public void drawAllLeftJustTextureShape(boolean topfirst)
 	{
 		int u = this.textureU;
